requirements.txt
pdfplumber>=0.10.0
python-pptx>=0.6.23
ollama>=0.3.0
streamlit>=1.35.0
